build_test, DWYU, include-cleaner for tombstone_receiver
Bug: b/523396865
Bug: b/532697459
Bug: b/534499070

+++ base/cvd/cuttlefish/host/commands/tombstone_receiver/main.cpp
@@ -14,12 +14,18 @@
  * limitations under the License.
  */
 
+#include <stddef.h>
+#include <stdint.h>
+#include <sys/time.h>  // IWYU pragma: keep: struct timeval
+
 #include <chrono>
 #include <fstream>
+#include <string>
+#include <vector>
 
 #include "absl/log/check.h"
 #include "absl/log/log.h"
-#include "fmt/chrono.h"
+#include "fmt/chrono.h"  // IWYU pragma: keep
 #include "fmt/format.h"
 
 #include "cuttlefish/common/libs/fs/fd.h"
@@ -29,10 +35,11 @@
 #include "cuttlefish/flag_parser/gflags_compat.h"
 #include "cuttlefish/flag_parser/shared_fd_flag.h"
 #include "cuttlefish/host/libs/config/logging.h"
+#include "cuttlefish/result/result_type.h"
 
 namespace cuttlefish {
 
-static uint num_tombstones_in_last_second = 0;
+static uint64_t num_tombstones_in_last_second = 0;
 static std::string last_tombstone_name = "";
 
 static std::string next_tombstone_path(const std::string& dir) {
@@ -95,6 +102,7 @@ int TombstoneReceiverMain(int argc, char** argv) {
     SharedFDSet error_set;
     error_set.Set(conn);
     while (file.is_open()) {
+      // NOLINTNEXTLINE(misc-include-cleaner): struct timeval
       struct timeval timeout = {TIMEOUT_SEC, 0};
       auto val = Select(&read_set, nullptr, &error_set, &timeout);
       if (val == 0) {
